backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
websockets==15.0.1
argon2-cffi==25.1.0
itsdangerous==2.2.0
pydantic==2.11.7
httpx==0.28.1
pytest==8.4.1
